MIR-505, MIR-508 Refactor (eckit::ResourceUsage and eckit::Timer to mir::trace::*, atlas::util::Earth to mir::util::Earth), static analyser

+++ src/mir/caching/CroppingCache.cc
@@ -33,7 +33,7 @@ CroppingCache::CroppingCache() :
 
 void CroppingCacheEntry::print(std::ostream& out) const {
     out << "CroppingCacheEntry[size=" << mapping_.size() << ",bbox=" << bbox_
-        << ",size=" << eckit::Bytes(sizeof(size_t) * mapping_.size()) << "]";
+        << ",size=" << eckit::Bytes(double(sizeof(size_t) * mapping_.size())) << "]";
 }
