+++ src/mir/caching/matrix/FileLoader.cc
@@ -39,7 +39,7 @@ FileLoader::FileLoader(const std::string& name, const eckit::PathName& path) :
 FileLoader::~FileLoader() = default;
 
 void FileLoader::print(std::ostream& out) const {
-    out << "FileLoader[path=" << path_ << ",size=" << eckit::Bytes(buffer_.size()) << "]";
+    out << "FileLoader[path=" << path_ << ",size=" << eckit::Bytes(double(buffer_.size())) << "]";
 }
 
 const void* FileLoader::address() const {
